+++ src/io/parsers/ab1.js
@@ -1,5 +1,5 @@
 // just a pared down version of: https://github.com/eamitchell/ab1ToJSON/blob/master/clientside.js
-import shortid from "shortid";
+import randomid from "randomid";
 
 const tagDict = {
   baseCalls2: { tagName: "PBAS", tagNum: 2, typeToReturn: "getChar" },
@@ -136,7 +136,7 @@ export default (file, name) => {
   };
 
   // convert to an aligned part format
-  const id = shortid.generate();
+  const id = randomid();
   const seq = baseCalls.join("");
   return { id, name, seq, traces };
 };