+++ src/utils/isEqual.js
@@ -1,0 +1,31 @@
+/**
+ * Deep equality comparison between two objects
+ *
+ * copy-paste from StackOverflow: https://stackoverflow.com/a/25456134
+ *
+ * @param {object} x left object for comparison
+ * @param {object} y right object for comparison
+ */
+const isEqual = (x, y) => {
+  if (x === y) {
+    return true;
+  } else if (
+    typeof x === "object" &&
+    x !== null &&
+    typeof y === "object" &&
+    y !== null
+  ) {
+    if (Object.keys(x).length !== Object.keys(y).length) return false;
+
+    for (var prop in x) {
+      if (y.hasOwnProperty(prop)) {
+        if (!isEqual(x[prop], y[prop])) return false;
+      } else return false;
+    }
+
+    return true;
+  }
+  return false;
+};
+
+export default isEqual;